Refactored integration tests into own file
