+++ datahub/bed_api/tests/test_repositories.py
@@ -1,10 +1,12 @@
 from unittest import mock
 
 import pytest
-from simple_salesforce import format_soql
 
-from datahub.bed_api.constants import ContactQuery
-from datahub.bed_api.models import EditAccount, EditContact, EditEvent
+from datahub.bed_api.models import (
+    EditAccount,
+    EditContact,
+    EditEvent,
+)
 from datahub.bed_api.repositories import (
     AccountRepository,
     ContactRepository,
@@ -14,7 +16,6 @@ from datahub.bed_api.repositories import (
 from datahub.bed_api.tests.test_utils import (
     create_fail_query_response,
     create_success_query_response,
-    NOT_BED_INTEGRATION_TEST_READY,
 )
 
 
@@ -23,8 +24,8 @@ class TestSalesforceRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_query_calls_salesforce_query_with_valid_args(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test query_more calls Salesforce with the correct Arguments
@@ -49,8 +50,8 @@ class TestSalesforceRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_query_next_calls_salesforce_query_more_with_valid_args(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test query_more calls Salesforce with the correct Arguments
@@ -78,8 +79,8 @@ class TestSalesforceRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_add_throws_not_implemented_error_by_default(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test add throws NotImplementedError
@@ -92,8 +93,8 @@ class TestSalesforceRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_delete_throws_not_implemented_error_by_default(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test delete throws NotImplementedError
@@ -106,8 +107,8 @@ class TestSalesforceRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_exists_throws_not_implemented_error_by_default(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test exists throws NotImplementedError
@@ -120,8 +121,8 @@ class TestSalesforceRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_get_throws_not_implemented_error_by_default(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test exists throws NotImplementedError
@@ -134,8 +135,8 @@ class TestSalesforceRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_get_by_throws_not_implemented_error_by_default(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test get_by throws NotImplementedError
@@ -148,8 +149,8 @@ class TestSalesforceRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_update_throws_not_implemented_error_by_default(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test update throws NotImplementedError
@@ -317,9 +318,9 @@ class TestAccountRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_add_calls_salesforce_contact_add_with_valid_args(
-            self,
-            mock_salesforce,
-            generate_account: EditAccount,
+        self,
+        mock_salesforce,
+        generate_account: EditAccount,
     ):
         """
         Test add calls Salesforce with the correct Arguments
@@ -337,8 +338,8 @@ class TestAccountRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_delete_calls_salesforce_account_delete_with_valid_args(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test delete calls Salesforce with the correct Arguments
@@ -356,8 +357,8 @@ class TestAccountRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_exists_return_true_when_query_response_succeeds(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test exists calls Salesforce with the correct Arguments
@@ -380,8 +381,8 @@ class TestAccountRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_exists_return_false_when_query_response_fails(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test exists calls Salesforce with the correct Arguments
@@ -401,8 +402,8 @@ class TestAccountRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_get_calls_salesforce_account_get_with_valid_args(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test get calls Salesforce with the correct Arguments
@@ -420,8 +421,8 @@ class TestAccountRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_get_by_calls_salesforce_account_get_with_valid_args(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test get_by calls Salesforce with the correct Arguments
@@ -441,9 +442,9 @@ class TestAccountRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_update_calls_salesforce_account_update_with_valid_args(
-            self,
-            mock_salesforce,
-            generate_account: EditContact,
+        self,
+        mock_salesforce,
+        generate_account: EditContact,
     ):
         """
         Test add calls Salesforce with the correct Arguments
@@ -468,9 +469,9 @@ class TestEventRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_add_calls_salesforce_event_add_with_valid_args(
-            self,
-            mock_salesforce,
-            generate_event: EditEvent,
+        self,
+        mock_salesforce,
+        generate_event: EditEvent,
     ):
         """
         Test add calls Salesforce with the correct Arguments
@@ -488,8 +489,8 @@ class TestEventRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_delete_calls_salesforce_event_delete_with_valid_args(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test delete calls Salesforce with the correct Arguments
@@ -507,8 +508,8 @@ class TestEventRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_exists_return_true_when_query_response_succeeds(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test exists calls Salesforce with the correct Arguments
@@ -531,8 +532,8 @@ class TestEventRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_exists_return_false_when_query_response_fails(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test exists calls Salesforce with the correct Arguments
@@ -552,8 +553,8 @@ class TestEventRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_get_calls_salesforce_event_get_with_valid_args(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test get calls Salesforce with the correct Arguments
@@ -571,8 +572,8 @@ class TestEventRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_get_by_calls_salesforce_event_get_with_valid_args(
-            self,
-            mock_salesforce,
+        self,
+        mock_salesforce,
     ):
         """
         Test get_by calls Salesforce with the correct Arguments
@@ -592,9 +593,9 @@ class TestEventRepositoryShould:
 
     @mock.patch('datahub.bed_api.factories.Salesforce')
     def test_update_calls_salesforce_event_update_with_valid_args(
-            self,
-            mock_salesforce,
-            generate_event: EditEvent,
+        self,
+        mock_salesforce,
+        generate_event: EditEvent,
     ):
         """
         Test update calls Salesforce with the correct Arguments
@@ -612,285 +613,3 @@ class TestEventRepositoryShould:
             'test_record_id',
             generate_event.as_values_only_dict(),
         )
-
-
-@pytest.mark.salesforce_test
-@pytest.mark.skipif(
-    NOT_BED_INTEGRATION_TEST_READY,
-    reason='BED security configuration missing from env file',
-)
-class TestIntegrationEventRepositoryShould:
-    """
-    Integration Test Event Repositories as Event
-    NOTE: Integration Tests needing BED configuration within
-    .env - see Vault for valid sandbox only settings
-        BED_USERNAME
-        BED_PASSWORD
-        BED_SECURITY_TOKEN
-        BED_IS_SANDBOX
-    """
-
-    def test_create_an_event_utilising(
-            self,
-            event_repository,
-            faker,
-            generate_event: EditEvent,
-    ):
-        """
-        Test BedFactory integration with the contact and account repositories
-        sampling all functions in an idempotent way, generating potential test
-        data for unit tests
-        :param event_repository: EventRepository fixture
-        :param faker: Faker library for generating data
-        :param generate_event: New event record generated with faker data
-        """
-        try:
-            # Create and update an event checking data and that the record exists
-            event_id = self.create_and_assert_event(event_repository, generate_event)
-
-            assert event_id is not None
-        finally:
-            # Delete and check event exists
-            if event_id:
-                assert event_id is not None
-
-    def create_and_assert_event(
-        self,
-        event_repository,
-        event: EditEvent,
-    ):
-        """
-        Create event on Salesforce and validate the data passed is generated as expected
-        :param event_repository: EventRepository fixture
-        :param event: New event record generated with faker data
-        :return: New event id
-        """
-        event_response = event_repository.add(
-            event.as_values_only_dict(),
-        )
-        assert event_response is not None
-        assert event_response['success'] is True
-        event_id = event_response['id']
-        assert event_id is not None
-        event.Id = event_id
-        self.assert_all_event_data(event, event_repository)
-        return event_id
-
-    def assert_all_event_data(self, event, event_repository):
-        """
-        Verify all data added onto BED or Salesforce and check the record exists
-        :param event: New event record generated with faker data
-        :param event_repository:
-        :param event_repository: EventRepository fixture
-        """
-        event_exists = event_repository.exists(event.Id)
-        assert event_exists is True
-        event_data = event_repository.get(event.Id)
-        for key, value in event.as_values_only_dict().items():
-            assert event_data[key] == value
-
-
-@pytest.mark.salesforce_test
-@pytest.mark.skipif(
-    NOT_BED_INTEGRATION_TEST_READY,
-    reason='BED security configuration missing from env file',
-)
-class TestIntegrationContactWithAccountRepositoryShould:
-    """
-    Integration Test Contact and Account Repositories as Contact is dependent on an Account
-    NOTE: Integration Tests needing BED configuration within
-    .env - see Vault for valid sandbox only settings
-        BED_USERNAME
-        BED_PASSWORD
-        BED_SECURITY_TOKEN
-        BED_IS_SANDBOX
-    """
-
-    def test_create_an_account_with_contact_utilising(
-            self,
-            contact_repository,
-            account_repository,
-            faker,
-            generate_account: EditAccount,
-            generate_contact: EditContact,
-    ):
-        """
-        Test BedFactory integration with the contact and account repositories
-        sampling all functions in an idempotent way, generating potential test
-        data for unit tests
-        :param contact_repository: ContactRepository fixture
-        :param account_repository: AccountRepository fixture
-        :param faker: Faker library for generating data
-        :param generate_account: New account record generated with faker data
-        :param generate_contact: New contact record generated with faker data
-        """
-        new_contact_id = None
-        new_account_id = None
-        try:
-            # Create a new account / organization / company
-            new_account_id = self.generate_and_assert_account(
-                account_repository,
-                generate_account,
-            )
-
-            # Create contact
-            generate_contact.AccountId = new_account_id
-            new_contact_id = self.generate_and_assert_contact(
-                contact_repository,
-                generate_contact,
-            )
-
-            #  Update Contact
-            self.update_and_assert_contact(contact_repository, new_contact_id, faker)
-
-            # TODO Update Account
-
-        finally:
-            #  Clean up generated data
-            if new_contact_id:
-                self.delete_and_assert_contact_deletion(
-                    contact_repository,
-                    new_contact_id,
-                )
-            if new_account_id:
-                self.delete_and_assert_account_deletion(
-                    account_repository,
-                    new_account_id,
-                )
-
-    def delete_and_assert_account_deletion(
-            self,
-            account_repository,
-            account_id,
-    ):
-        """
-        Delete generated account from the database
-        :param account_repository: AccountRepository fixture
-        :param account_id: Account id to delete
-        """
-        delete_account_response = account_repository.delete(account_id)
-        assert delete_account_response is not None
-        assert delete_account_response == 204
-        exists = account_repository.exists(account_id)
-        assert exists is False
-
-    def delete_and_assert_contact_deletion(
-            self,
-            contact_repository,
-            contact_id,
-    ):
-        """
-        Delete generated contact from the database
-        :param contact_repository: ContactRepository fixture
-        :param contact_id: Contact id to delete
-        """
-        delete_contact_response = contact_repository.delete(contact_id)
-        assert delete_contact_response is not None
-        assert delete_contact_response == 204
-        exists = contact_repository.exists(contact_id)
-        assert exists is False
-
-    def update_and_assert_contact(
-            self,
-            contact_repository,
-            new_contact_id,
-            faker,
-    ):
-        """
-        Update the contact with basic notes testing update
-        :param contact_repository: ContactRepository fixture
-        :param new_contact_id: Contact id to update
-        :param faker: Faker library
-        """
-        # Example using original edit object sending all values
-        # contact.Notes__c = 'Integration Test Notes - Update'
-        # update_contact_response = contact_repository.update(
-        #     f'Id/{new_contact_id}', contact.as_values_only_dict())
-        notes_update = faker.text(max_nb_chars=100)
-        update_contact_response = contact_repository.update(
-            f'Id/{new_contact_id}',
-            dict(Notes__c=notes_update),
-        )
-        assert update_contact_response is not None
-        assert update_contact_response == 204
-
-        contact_check = contact_repository.query(
-            format_soql(
-                ContactQuery.get_notes_by_id.value.sql,
-                id=new_contact_id,
-            ),
-        )
-        assert contact_check is not None
-        assert contact_check['totalSize'] == 1
-        assert contact_check['done'] is True
-        assert contact_check['records'][0]['Id'] == new_contact_id
-        assert contact_check['records'][0]['Notes__c'] == notes_update
-
-    def generate_and_assert_contact(
-            self,
-            contact_repository,
-            contact: EditContact,
-    ):
-        """
-        Create Account data on Salesforce testing as many ContactRepository
-        Methods as possible
-        :param contact_repository: ContactRepository fixture
-        :param contact: Random Generated Contact
-        :return: new contact id
-        """
-        contact_add_response = contact_repository.add(contact.as_values_only_dict())
-        assert contact_add_response is not None
-        assert contact_add_response['success'] is True
-        contact_id = contact_add_response['id']
-        contact.Id = contact_id
-        self.assert_all_contact_data(contact, contact_repository)
-        return contact_id
-
-    def assert_all_contact_data(self, contact, contact_repository):
-        """
-        Verify contact exists on Salesforce and the data is the same as the
-        data used to generate teh contact
-        :param contact: Random Generated Contact
-        :param contact_repository: ContactRepository fixture
-        """
-        contact_exists = contact_repository.exists(contact.Id)
-        assert contact_exists is True
-        contact_data = contact_repository.get_by(
-            'Datahub_ID__c',
-            contact.Datahub_ID__c,
-        )
-        for key, value in contact.as_values_only_dict().items():
-            assert contact_data[key] == value
-
-    def generate_and_assert_account(
-            self,
-            account_repository,
-            account: EditAccount,
-    ):
-        """
-        Create Account Data on Salesforce using dynamic data
-        :param account_repository: AccountRepository fixture
-        :param account: New account record generated with faker data
-        :return: Account Id
-        """
-        account_add_response = account_repository.add(account.as_values_only_dict())
-        assert account_add_response is not None
-        assert account_add_response['success'] is True
-        account_id = account_add_response['id']
-        assert account_id is not None
-        account.Id = account_id
-        self.assert_all_account_data(account, account_repository)
-        return account_id
-
-    def assert_all_account_data(self, account, account_repository):
-        """
-        Verify all data added onto BED or Salesforce and check the record exists
-        :param account: New account record generated with faker data
-        :param account_repository:
-        :param account_repository: AccountRepository fixture
-        """
-        account_exists = account_repository.exists(account.Id)
-        assert account_exists is True
-        account_data = account_repository.get(account.Id)
-        for key, value in account.as_values_only_dict().items():
-            assert account_data[key] == value
